+++ lib/screens/widgets/help_section.dart
